Add theme management commands
Introduces the ability to create, edit and remove templates from which
themes are derived.

Extension activation is limited to explicitly executing available
commands or if a JSON file is being edited. In practice, once a user
is happy with their theme configuration, they should not need to incur
the cost of activating this extension when all it does is contribute
themes.